backend/requirements.txt
django
djangorestframework
django-cors-headers
requests
yfinance
pandas
python-dotenv
gunicorn
whitenoise
google-auth
requests-oauthlib
djangorestframework-simplejwt
dj_database_url
psycopg2
redis
celery
pydantic
feedparser>=6.0.10
